+++ models/photo.go
@@ -174,7 +174,7 @@ func (m *Photo) GetFieldDeserializers()(map[string]func(i878a80d2330e89d26896388
         return nil
     }
     res["orientation"] = func (n i878a80d2330e89d26896388a3f487eef27b0a0e6c010c493bf80be1452208f91.ParseNode) error {
-        val, err := n.GetInt32Value()
+        val, err := n.GetFloat64Value()
         if err != nil {
             return err
         }
@@ -244,14 +244,14 @@ func (m *Photo) GetOdataType()(*string) {
     return nil
 }
 // GetOrientation gets the orientation property value. The orientation value from the camera. Writable on OneDrive Personal.
-// returns a *int32 when successful
-func (m *Photo) GetOrientation()(*int32) {
+// returns a *float64 when successful
+func (m *Photo) GetOrientation()(*float64) {
     val, err := m.GetBackingStore().Get("orientation")
     if err != nil {
         panic(err)
     }
     if val != nil {
-        return val.(*int32)
+        return val.(*float64)
     }
     return nil
 }
@@ -318,7 +318,7 @@ func (m *Photo) Serialize(writer i878a80d2330e89d26896388a3f487eef27b0a0e6c010c4
         }
     }
     {
-        err := writer.WriteInt32Value("orientation", m.GetOrientation())
+        err := writer.WriteFloat64Value("orientation", m.GetOrientation())
         if err != nil {
             return err
         }
@@ -405,7 +405,7 @@ func (m *Photo) SetOdataType(value *string)() {
     }
 }
 // SetOrientation sets the orientation property value. The orientation value from the camera. Writable on OneDrive Personal.
-func (m *Photo) SetOrientation(value *int32)() {
+func (m *Photo) SetOrientation(value *float64)() {
     err := m.GetBackingStore().Set("orientation", value)
     if err != nil {
         panic(err)
@@ -431,7 +431,7 @@ type Photoable interface {
     GetFocalLength()(*float64)
     GetIso()(*int32)
     GetOdataType()(*string)
-    GetOrientation()(*int32)
+    GetOrientation()(*float64)
     GetTakenDateTime()(*i336074805fc853987abe6f7fe3ad97a6a6f3077a16391fec744f671a015fbd7e.Time)
     SetBackingStore(value ie8677ce2c7e1b4c22e9c3827ecd078d41185424dd9eeb92b7d971ed2d49a392e.BackingStore)()
     SetCameraMake(value *string)()
@@ -442,6 +442,6 @@ type Photoable interface {
     SetFocalLength(value *float64)()
     SetIso(value *int32)()
     SetOdataType(value *string)()
-    SetOrientation(value *int32)()
+    SetOrientation(value *float64)()
     SetTakenDateTime(value *i336074805fc853987abe6f7fe3ad97a6a6f3077a16391fec744f671a015fbd7e.Time)()
 }
